Updated terraform-google-provider to 6.15.0
This requires deletion of the removed resource
`google_identity_platform_project_default_config`.

This also requires a bump of the dependency of
`github.com/hashicorp/terraform-provider-google` in `go.mod` to 6.x.y but
it is explicitly setup as a invalid Go module. See
https://developer.hashicorp.com/terraform/plugin/best-practices/interacting-with-providers#do-not-import-providers-as-go-modules
To workaroud this the provider code was checked out locally and a
`replace` statement is used to use that code.